% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt.R
\name{power_test}
\alias{power_test}
\title{Função para o cálculo do poder do teste}
\usage{
power_test(
  N = 1000L,
  B = 250L,
  n,
  f,
  sig,
  q,
  kicks,
  par0,
  ncores,
  chisinf = FALSE,
  ...
)
}
\description{
Função para o cálculo do poder do teste
}
\examples{
pdf_ew <- function(par, x, var = NULL){
  alpha <- par[1]
  sigma <- par[2]
  theta <- par[3]

if (is.list(var)) eval(parse(text = paste(var[[1]], " <- ", unlist(var[[2]]), sep = "")))

  alpha * theta / sigma * (1 - exp(-(x / sigma) ^ alpha)) ^ (theta - 1) *
    exp(-(x / sigma) ^ alpha) * (x / sigma) ^ (alpha - 1)
}

rew <- function(n, alpha, sigma, theta){
  u <- runif(n, 0, 1)
  sigma * (-log(1 - u ^ (1 / theta))) ^ (1 / alpha)
}

set.seed(1L, kind = "L'Ecuyer-CMRG")

tictoc::tic()
power_test(N = 200L,
           B = 250L,
           n = 250L,
           f = pdf_ew,
           sig = 0.05,
           q = rew,
           kicks = c(1, 1, 1),
           par0 = list(c("sigma", "theta"), c(1.5, 1.8)),
           ncores = 4L,
           alpha = 1.2,
           sigma = 1,
           theta = 1.3
)
tictoc::toc()
}
